requirements.txt
kubernetes
streamlit
faiss-cpu
sentence-transformers
requests
python-dotenv
matplotlib
numpy
pandas